+++ ChangeLog.md
@@ -7,6 +7,7 @@
 * New Option MaxPDULength in DicomServiceOption.
 * Rewrite SOP class extended negotiation support (#740)
 * Bub fix: extended negotiations have been accepted by default, but should be explicitly accepted (#900)
+* Global static property DicomValidation.PerformValidation to turn off validation for every DicomDataset. 
 
 #### v.4.0.2 (7/30/2019)
 * Bug fix: prevent resource leak when DesktopNetworkListener waits for new TCP clients